writeloader: make it work with recent kernel
The tools does not work with newer kernel, see:

http://lists.infradead.org/pipermail/linux-mtd/2013-December/050739.html

This patch fixes the issue and also changes the main loop to be a little
bit more efficient and readable.

Original:
   text    data     bss     dec     hex filename
   4554     368      20    4942    134e writeloader
Modified:
   text    data     bss     dec     hex filename
   4430     368      20    4818    12d2 writeloader

It also calls closedir on /sys/bus/platform/devices/ that was not
done and removes sizeof(unsigned char) since is 1 by definition.

Signed-off-by: Ladislav Michl <ladis@linux-mips.org>